Revert "Pull relevant parts of 7e56613"
This seems to require previous commits that were initially thought to be only relevant for chromium browsers. It will have to be investigated some other time if this or anything related to this that should be pulled over. This reverts commit 028d5451e85a5c427ffa8cde980cae56d82e8845.